% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpart.R
\name{autoplot.rpart}
\alias{autoplot.rpart}
\title{Creates ggplot from \link[rpart]{rpart.object}}
\usage{
\method{autoplot}{rpart}(object, ...)
}
\arguments{
\item{object}{an \link[rpart]{rpart.object}}

\item{...}{additional parameters of function \link[survival.tzh]{autoplot.survfit}}
}
\value{
Function \code{\link[=autoplot.rpart]{autoplot.rpart()}} returns a \link[ggplot2]{ggplot} object.
}
\description{
..
}
\keyword{internal}
