% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfit.rpart.R
\name{autoplot.survfit.rpart}
\alias{autoplot.survfit.rpart}
\title{\code{\link[=autoplot.survfit.rpart]{autoplot.survfit.rpart()}}}
\usage{
\method{autoplot}{survfit.rpart}(object, ...)
}
\arguments{
\item{object}{returned value of function \code{\link[=survfit.rpart]{survfit.rpart()}}}

\item{...}{..}
}
\description{
..
}
\keyword{internal}
