% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsurvplot_rpart.R
\name{ggsurvplot_rpart}
\alias{ggsurvplot_rpart}
\title{Kaplan-Meier Curves of \link[rpart]{rpart} via \CRANpkg{survminer}}
\usage{
ggsurvplot_rpart(fit, ...)
}
\arguments{
\item{fit}{\link[rpart]{rpart}}

\item{...}{..}
}
\description{
Kaplan-Meier curves based on recursive partitioning and regression trees \link[rpart]{rpart}.
}
\examples{
library(survival)
library(rpart)
rp = rpart(Surv(time, status) ~ age, data = veteran, maxdepth = 2L, model = TRUE)
ggsurvplot_rpart(rp)

}
