% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpart.R
\name{is.Surv.endpoint.rpart}
\alias{is.Surv.endpoint.rpart}
\title{Does an \link[rpart]{rpart.object} Has a \link[survival]{Surv} Endpoint?}
\usage{
\method{is.Surv.endpoint}{rpart}(object)
}
\arguments{
\item{object}{an \link[rpart]{rpart.object}}
}
\description{
Does an \link[rpart]{rpart.object} Has a \link[survival]{Surv} Endpoint?
}
\keyword{internal}
