% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{nobs.rpart}
\alias{nobs.rpart}
\title{Sample Size in \link[rpart]{rpart} Object}
\usage{
\method{nobs}{rpart}(object, ...)
}
\arguments{
\item{object}{\link[rpart]{rpart} object}

\item{...}{additional parameters, not currently in use}
}
\value{
Function \link{nobs.rpart} returns an \link[base]{integer} scalar
}
\description{
Method dispatch for S3 generic \link[stats]{nobs}.
}
