% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risklev.R
\name{risklev}
\alias{risklev}
\title{Risk Level of \link[rpart]{rpart} Object}
\usage{
risklev(object, ...)
}
\arguments{
\item{object}{\link[rpart]{rpart}}

\item{...}{..}
}
\value{
Function \link{risklev} returns a \link[base]{factor}.
}
\description{
..
}
\note{
Function \link[rpart]{labels.rpart} (for S3 generic \link[base]{labels}) only captures the last branching information, which is not ideal.
}
\examples{
library(rpart)
rp = rpart(Kyphosis ~ Age + Number + Start, data = kyphosis, model = TRUE)
rp |> prp_()
rp |> risklev()
rp |> ggsurvplot_rpart()

}
