% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpart.R
\name{rpart}
\alias{rpart}
\title{\link[rpart]{rpart.object}}
\description{
\link[rpart]{rpart.object}
}
\examples{
library(survival)
vet = survival::veteran |>
 within.data.frame(expr = {
  time = as.difftime(time, units = 'days')
  os = Surv(time, status)
 })

list(
 'non survival' = rpart(Kyphosis ~ Age + Start, data = kyphosis, model = TRUE),
 'survival' = rpart(os ~ age, data = vet, maxdepth = 2L, model = TRUE)
) |> fastmd::render2html()

}
