% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{rpart.tzh-package}
\alias{rpart.tzh}
\alias{rpart.tzh-package}
\title{rpart.tzh: Extensions for rpart}
\description{
To extend rpart package.
}
\note{
Help files of individual functions are intentionally suppressed in the \code{pdf} manual.
Users are encouraged to get started with the vignettes.
}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com}

}
