% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpart.R
\name{survdiff_.rpart}
\alias{survdiff_.rpart}
\title{Test Survival Curve Differences of \link[rpart]{rpart.object}}
\usage{
\method{survdiff_}{rpart}(object, fmt = "\%.2g", ...)
}
\arguments{
\item{object}{an \link[rpart]{rpart.object}}

\item{fmt}{\link[base]{character} scalar,
string formatting of the leaf-risk,
see function \link[base]{sprintf}}

\item{...}{additional parameters of function \link[survival]{survdiff}}
}
\description{
Test survival curve differences of an \link[rpart]{rpart.object},
if the endpoint is a \link[survival]{Surv} object.
}
\examples{
library(survival.tzh)
rpart(Surv(time, status) ~ age, data = veteran, maxdepth = 2L, model = TRUE) |> 
 survdiff_()

}
