% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpart.R
\name{survfit.rpart}
\alias{survfit.rpart}
\title{Kaplan-Meier Curves of \link[rpart]{rpart.object}}
\usage{
\method{survfit}{rpart}(formula, fmt = "\%.2g", ...)
}
\arguments{
\item{formula}{an \link[rpart]{rpart.object}}

\item{fmt}{\link[base]{character} scalar,
string formatting of the leaf-risk,
see function \link[base]{sprintf}}

\item{...}{additional parameters of function \link[survival]{survfit.formula}}
}
\value{
The \code{S3} method \code{\link[=survfit.rpart]{survfit.rpart()}} returns a \link[survival]{survfit.object}.
}
\description{
Kaplan-Meier curves of an \link[rpart]{rpart.object},
if the endpoint is a \link[survival]{Surv} object.
}
\examples{
library(survival)
rpart(Surv(time, status) ~ age, data = veteran, maxdepth = 2L, model = TRUE) |> 
  survfit()

}
\keyword{internal}
