% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_.rpart.R
\name{md_.rpart}
\alias{md_.rpart}
\title{Create Markdown Lines for \link[rpart]{rpart} Objects}
\usage{
\method{md_}{rpart}(x, ...)
}
\arguments{
\item{x}{\link[rpart]{rpart} object}

\item{...}{..}
}
\value{
Function \code{\link[=md_.rpart]{md_.rpart()}} returns a \link[base]{character} \link[base]{vector}.
}
\description{
..
}
\examples{
list(
 '`rpart`' = rpart(Kyphosis ~ Age + Number + Start, data = kyphosis, model = TRUE)
) |> rmd.tzh::render_(file = 'rpart')

}
\keyword{internal}
