% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_.rpart.R
\name{md_.rpart}
\alias{md_.rpart}
\title{Create Markdown Lines for \link[rpart]{rpart} Objects}
\usage{
md_.rpart(x, xnm, ...)
}
\arguments{
\item{x}{\link[rpart]{rpart} object}

\item{xnm}{..}

\item{...}{..}
}
\value{
Function \code{\link[=md_.rpart]{md_.rpart()}} returns a \link[base]{character} \link[base]{vector}.
}
\description{
..
}
\keyword{internal}
