% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpart.R
\name{nobs.rpart}
\alias{nobs.rpart}
\title{Sample Size in \link[rpart]{rpart.object}}
\usage{
\method{nobs}{rpart}(object, ...)
}
\arguments{
\item{object}{an \link[rpart]{rpart.object}}

\item{...}{additional parameters, not currently in use}
}
\value{
Function \code{\link[=nobs.rpart]{nobs.rpart()}} returns an \link[base]{integer} scalar.
}
\description{
Method dispatch for S3 generic \link[stats]{nobs}.
}
\keyword{internal}
