% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prp_.R
\name{prp_}
\alias{prp_}
\title{My favorite customization of \link[rpart.plot]{prp}}
\usage{
prp_(x, ...)
}
\arguments{
\item{x}{\link[rpart]{rpart} object}

\item{...}{additional parameters, currently of no use}
}
\value{
Function \code{\link[=prp_]{prp_()}} does not have a returned value.
}
\description{
Do NOT overwrites function \code{rpart:::print.rpart()}.
}
\note{
Function \code{ggdendro::ggdendrogram} will plot \link[rpart]{rpart} object, but not very good.

Another option is \code{?partykit::plot.party}, which seems to requires numeric endpoint and/or branching criterion (?)
}
\references{
\url{http://www.di.fc.ul.pt/~jpn/r/tree/tree.html}

Inspired by function \code{heat.tree()} in \url{http://www.milbo.org/rpart-plot/prp.pdf}
}
\keyword{internal}
