% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpart.R
\name{survfit.rpart}
\alias{survfit.rpart}
\title{Survival Curves based on \link[rpart]{rpart}}
\usage{
\method{survfit}{rpart}(formula, fmt = "\%.2g", ...)
}
\arguments{
\item{formula}{\link[rpart]{rpart}}

\item{fmt}{\link[base]{character} scalar}

\item{...}{..}
}
\value{
Function \code{\link[=survfit.rpart]{survfit.rpart()}} returns a \code{'survfit.rpart'} object,
an derived class of \code{S3} class \link[survival]{survfit}.
}
\description{
..
}
\examples{
library(survival)
rpart(Surv(time, status) ~ age, data = veteran, maxdepth = 2L, model = TRUE) |> 
  survfit()

}
\keyword{internal}
